Cleansed solvespace.h, gl3shader
