+++ src/render/render.h
@@ -23,29 +23,7 @@ typedef struct _cairo_surface cairo_surface_t;
 
 namespace SolveSpace {
 
-// A description of scene lighting.
-class Lighting {
-public:
-    RgbaColor   backgroundColor;
-    double      ambientIntensity;
-    double      lightIntensity[2];
-    Vector      lightDirection[2];
-};
-
 class BatchCanvas;
-
-// An interface for view-dependent visualization.
-class ViewportCanvas : public Canvas {
-public:
-    virtual void SetCamera(const Camera &camera) = 0;
-    virtual void SetLighting(const Lighting &lighting) = 0;
-
-    virtual void NewFrame() = 0;
-    virtual void FlushFrame() = 0;
-    virtual std::shared_ptr<Pixmap> ReadFrame() = 0;
-
-    virtual void GetIdent(const char **vendor, const char **renderer, const char **version) = 0;
-};
 
 // An interface for view-independent visualization.
 class BatchCanvas : public Canvas {
